routes route and component created, route names listed
